Seperated guiding statistics classes into own module
