+++ pyobs/modules/pointing/guidingstatistics/__init__.py
@@ -1,0 +1,2 @@
+from .uptime import GuidingStatisticsUptime
+from .pixeloffset import GuidingStatisticsPixelOffset
